requirements.txt
requests
python-dotenv
groq
shodan
requests
beautifulsoup4
tkinter